Refactoring pipeline, creating misc and plotting
