+++ mothra/tests/test_pipeline.py
@@ -3,13 +3,7 @@ import os
 import shutil
 import pytest
 
-import pipeline
-
-from csv import reader
 from glob import glob
-from pathlib import Path
-from skimage.io import imread
-from skimage.util import img_as_float
 
 
 PATH_TEST_FILES = 'mothra/tests/test_files'
@@ -51,105 +45,3 @@ def test_pipeline_main():
 
     # assert the two outputs exist
     assert(output_image and output_csv)
-
-
-def test_create_layout():
-    """Checks if axes for regular and detailed plots were created
-    properly.
-
-    Summary
-    -------
-    We pass different stages and plot levels to pipeline.create_layout,
-    and check the resulting axes.
-
-    Expected
-    --------
-    - pipeline.create_layout(1, 0) should not return axes.
-    - pipeline.create_layout(3, 2) should return all seven axes
-    (ax_main, ax_bin, ax_poi, ax_structure, ax_signal, ax_fourier, ax_tags).
-    - pipeline.create_layout(3, 1) should return a list with three axes and
-    four None.
-    """
-    axes = pipeline.create_layout(1, 0)
-    assert axes is None
-    axes = pipeline.create_layout(3, 2)
-    for ax in axes:
-        assert ax
-    axes = pipeline.create_layout(3, 1)
-    for ax in axes[:3]:
-        assert ax
-    for ax in axes[3:]:
-        assert ax is None
-
-
-def test_read_orientation():
-    """Checks if orientation is extracted correctly from EXIF data.
-
-    Summary
-    -------
-    We provide an input image with known angle and compare its angle
-    read by pipeline.read_orientation.
-
-    Expected
-    --------
-    Orientation for the input image is 6, (right, top);
-    pipeline.read_orientation should return angle equals 90 deg.
-    """
-    angle = pipeline.read_orientation(TEST_TILTED_IMAGE)
-
-    assert angle == 90
-
-
-def test_process_paths_in_input():
-    """Checks if files from a folder within an input text file are correctly
-    read.
-
-    Summary
-    -------
-    We provide the path for a text file containing the path for a folder, and
-    check if the filenames are read by pipeline._process_paths_in_input.
-
-    Expected
-    --------
-    TEST_INPUT_IMAGES and image_paths should contain the same filenames.
-    """
-    image_paths = pipeline._process_paths_in_input(TEST_INPUT_FILE)
-
-    assert image_paths.sort() == TEST_INPUT_IMAGES.sort()
-
-
-def test_read_paths_in_file():
-    """Checks if files from a folder within an input text file are correctly
-    read.
-
-    Summary
-    -------
-    We provide the path for a text file containing the path for a folder, and
-    check if the filenames are read by pipeline._read_paths_in_file.
-
-    Expected
-    --------
-    TEST_INPUT_IMAGES and image_paths should contain the same filenames.
-    """
-    image_paths = pipeline._read_paths_in_file(TEST_INPUT_FILE)
-
-    assert image_paths.sort() == TEST_INPUT_IMAGES.sort()
-
-
-def test_read_filenames_in_folder():
-    """Check if filenames in folder are read properly.
-
-    Summary
-    -------
-    We pass a folder containing a number of known input images, and
-    check if pipeline._read_filenames_in_folders reads their filenames
-    properly.
-
-    Expected
-    --------
-    result_fnames and TEST_INPUT_IMAGES contain the same filenames.
-    """
-    test_folder = f'{PATH_TEST_FILES}/test_input/'
-    result_fnames = pipeline._read_filenames_in_folder(test_folder)
-
-    assert result_fnames.sort() == TEST_INPUT_IMAGES.sort()
